Update security-secrets-setup package name under internal to be consistent with other core services #1917
