+++ internal/security/secrets/certificates/interface_test.go
@@ -18,12 +18,12 @@ import (
 	"os"
 	"testing"
 
-	"github.com/edgexfoundry/edgex-go/internal/security/setup"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets"
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/logger"
 )
 
 func TestNewCertificateGenerator(t *testing.T) {
-	seed := setup.CertificateSeed{}
+	seed := secrets.CertificateSeed{}
 	writer := mockFileWriter{}
 	mockLogger := logger.MockLogger{}
