+++ internal/security/secrets/certificates/testdata/EdgeXTrustCA/EdgeXTrustCA.pem
