+++ internal/security/secrets/certificates/testdata/EdgeXTrustCA/EdgeXTrustCA.priv.key
