+++ internal/security/secrets/certificates/testdata/EdgeXTrustCA/EdgeXTrustCAInvalid.pem
