+++ internal/security/secrets/certificates/testdata/EdgeXTrustCA/EdgeXTrustCAInvalid.priv.key
