+++ internal/security/secrets/certificates/tls.go
@@ -25,14 +25,14 @@ import (
 	"math/big"
 	"time"
 
-	"github.com/edgexfoundry/edgex-go/internal/security/setup"
+	"github.com/edgexfoundry/edgex-go/internal/security/secrets"
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/logger"
 	"github.com/pkg/errors"
 )
 
 type tlsCertGenerator struct {
 	logger logger.LoggingClient
-	seed   setup.CertificateSeed
+	seed   secrets.CertificateSeed
 	writer FileWriter
 }
