+++ internal/security/secrets/init.go
@@ -12,7 +12,7 @@
  * the License.
  *******************************************************************************/
 
-package setup
+package secrets
 
 import (
 	"fmt"
