+++ internal/security/secrets/mocks/DirectoryHandler.go
