+++ internal/security/secrets/mocks/custom.go
