+++ internal/security/secrets/option/exitCode.go
