+++ internal/security/secrets/option/pkiOption.go
