+++ internal/security/secrets/option/pkiOption_test.go
