+++ internal/security/secrets/option/testdata/res/configuration.toml
