+++ internal/security/secrets/option/testdata/res/pkisetup-kong.json
