+++ internal/security/secrets/option/testdata/res/pkisetup-vault.json
